+++ src/ts/plugins/background.ts
@@ -2,29 +2,34 @@
  * 检测背景图片插件是否在运行
  * @param times 运行次数
  * @since 1.3.5
- * @version 2.5.2
+ * @version 2.6.1
  */
 export function bg(times: number) {
     // 背景自定义插件，部分情况下插件加载缓慢可重复检测一次
     const bglayer = document.getElementById("bglayer");
     if (bglayer) {
         const style = window.getComputedStyle(bglayer);
-        const body = document.body;
         if (style.getPropertyValue("display") !== "none") {
-            body.classList.add("bgenable");
+            document.body.classList.add("bgenable");
         } else if (style.getPropertyValue("display") === "none") {
             // console.log("disable background");
-            body.classList.remove("bgenable");
+            document.body.classList.remove("bgenable");
         }
         // 刚开始每2秒重新检测状态，检测10秒
         if (times < 5) {
-            globalThis.vscTimer.bgTimer = setTimeout(bg, 2000, times + 1);
+            globalThis.vscTimer.bgTimer = window.setTimeout(bg, 2000, times + 1);
         } else {
             globalThis.vscTimer.bgTimer = null;
         }
-    } else if (times === 0 || times === 1) {
+        if (globalThis.vscObserver.bgExistObserver === null) {
+            bgExistObserver();
+        }
+    } else if ((times === 0 || times === 1) && !bglayer) {
         // 未启用插件3秒后重新检测两遍
         setTimeout(bg, 3000, times + 1);
+    } else if (!bglayer) {
+        // 检测不到背景图移除属性
+        document.body.classList.remove("bgenable");
     }
 }
 
@@ -32,7 +37,7 @@ export function bg(times: number) {
  * 监听背景图片插件的属性修改
  * @param times 运行次数
  * @since 1.4.0
- * @version 2.4.2
+ * @version 2.6.1
  */
 export function bgobserver(times: number) {
     const bglayer = document.getElementById("bglayer");
@@ -52,12 +57,48 @@ export function bgobserver(times: number) {
         globalThis.vscTimer.bgObserTimer = null;
     } else {
         // if (times == 0 && !document.body.classList.contains('vscmobile')) {
-        if (times == 0) {
+        if (times === 0) {
             // 运行失败等待5秒
-            globalThis.vscTimer.bgObserTimer = setTimeout(bgobserver, 5000, 1);
-        } else if (times == 1) {
+            globalThis.vscTimer.bgObserTimer = window.setTimeout(bgobserver, 5000, 1);
+        } else if (times === 1) {
             console.error("背景插件监听失败，修改插件启用状态需手动刷新");
             globalThis.vscTimer.bgObserTimer = null;
         }
+    }
+}
+
+/**
+ * 检测背景是否被移除（插件关闭）
+ * @since 2.6.1
+ * @version 2.6.1
+ */
+function bgExistObserver() {
+    const bglayer = document.getElementById("bglayer");
+    let parent = null;
+    if (bglayer) {
+        parent = bglayer.parentElement;
+        globalThis.vscObserver.bgExistObserver = new MutationObserver(function (mutationsList) {
+            for (const mutation of mutationsList) {
+                if (mutation.type === "childList") {
+                    mutation.addedNodes.forEach((node) => {
+                        if (node.nodeType === Node.ELEMENT_NODE && (node as HTMLCanvasElement).id === "bglayer") {
+                            // 检测存在时添加属性
+                            bg(0);
+                            return;
+                        }
+                    });
+                    mutation.removedNodes.forEach((node) => {
+                        if (node.nodeType === node.ELEMENT_NODE && (node as HTMLCanvasElement).id === "bglayer") {
+                            // 检测不存在时直接移除属性
+                            bg(2);
+                            return;
+                        }
+                    });
+                }
+            }
+        });
+        globalThis.vscObserver.bgExistObserver.observe(parent, {
+            childList: true,
+        });
     }
 }